Fix latest batch of null warnings
2f79cfb was about fixing the errors that cropped up with the new
version of JDT. This latest patch fixes the related warnings.

It introduces external @NonNull annotations to the following
methods in Guava's Multimap:

get()
keySet()
values()

The last two follow the same reasoning as the corresponding Map
methods. The first one, get(), may seem surprising. But Multimap's
get() is really @NonNull, it will at most return an empty Collection
if there are no values in the map matching the passed key.

Change-Id: Ie8dccff9491a8c0200b503ad76fd44092eb975e2
Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>
Reviewed-on: https://git.eclipse.org/r/61929
Reviewed-by: Hudson CI
Reviewed-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Tested-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>