ctf: Add lost events to synthetic LTTng kernel trace generator
The trace generator now creates packets with lost events pseudo-randomly
approximately once out of 10 times the number of channels.

The trace generator also now sets the content size of the last packet of
each channel so that the total number of events (including lost events)
is exactly the number of requested trace events.

Change-Id: Ie804e5650d5a4186225b5dfcc75526b0fe9c2704
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/61599
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>