tmf: Use prioritized List instead of Set in XmlStateProvider
This allows to have a kind of priority between the event handlers. It
helps to improve dependency between the event handlers. Now, when
receiving an event E, it is possible that the result of a handler X
depends of the change made by a handler Y on the same event.

Change-Id: I33592e907d4c2f024a46cc365f0bbe796e940049
Signed-off-by: Jean-Christian Kouame <jean-christian.kouame@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/61672
Reviewed-by: Hudson CI
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>