analysis: add standard deviation to the segment store statistics
This patch will be useful for extracting more key metrics and flagging
outlyer segments.

Some reminders:

The standard deviation is the square root of the variance. The variance
is the sum of squared elements divided by the cardinality of a list. To
calculate the standard deviation incrementally, we keep an accumulator
of all the squares of a latency. This is then divided when needed by the
number of elements (segments) and square rooted.

Also, this patch calculates an online mean with less rounding errors.

Change-Id: Ia918f08f2351d7086bd05aac1ad645cfff13eb58
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/61824
Reviewed-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>