Remove API tooling from test plugins
We do not track the API of test plugins so it is not necessary to have
them configured for API tooling. This could lead to errors in the case
where the API baseline contained some test plugins. In Oomph, for
example, the API baseline is just a zip of the complete update site, so
it does contain the test plugins and this triggers errors.

Change-Id: I6bd735c2890ae5eaf329316d4f1bae43348c2281
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/61839
Reviewed-by: Hudson CI
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>