tmf: Add vertical zooming support to time graph viewer and combo
Vertical zoom is triggered with Ctrl+"+" and Ctrl+"-". Reset is
triggered by Ctrl+"0".

Ctrl+"=" is equivalent to Ctrl+"+" for vertical zoom-in, and "=" is now
equivalent to "+" for horizontal zoom-in also.

The font of the time graph control is changed along with the zooming,
and the views with post-draw handling to draw labels on time graph items
are modified to use a new method which centers the text vertically.

Change-Id: Ic75c1ebe7e937c612b64346e2814894c500f964f
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/62359
Reviewed-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>