Use maven-enforcer-plugin to enforce minimum versions
<prerequisites> tags are not inherited by children. We can use
maven-enforcer-plugin to really check the Maven version.

See http://stackoverflow.com/q/18055009/4227853 .

It's not really clear if <prerequisites> is really deprecated
or not, but "mvn versions" complains if it is not present, so
we can at least leave it in the top-level pom.

While at it we can also configure it to check for the minimum
Java version the project needs (1.8).

Change-Id: I607c33e3fecf39d932ae7a44d58b85ce20336dc7
Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>
Reviewed-on: https://git.eclipse.org/r/62280
Reviewed-by: Hudson CI
Reviewed-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Tested-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>