tmf: FilterColorEditorTest on Mac
On Mac, RGB values that are captured with ImageHelper are affected by
monitor color profiles. To account for this, we can draw the expected
color in a simple shell and use that color as expected value instead.

Also, switching to the Filter mode in the table is not instantaneous,
we need to wait until it's done before capturing the next image.

Change-Id: I930570027ba53f3dbdcecb5925a8e67c24c239b2
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/62573
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
(cherry picked from commit 7431c59e218960d40b952e5813ec6f3760e164bb)
Reviewed-on: https://git.eclipse.org/r/62683