releng: Fix alltests failure on Mac
The alltests plugin was failing on Mac because it tried to run 0 tests
on the UI thread but it needs to start on the first thread
(-XstartOnFirstThread). We can just set it to not un on the UI thread.
There is no need for the UI thread when running 0 tests.

Change-Id: I1e260f96e1ec51e256e8280262bcbd3c83a92de3
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/62574
Reviewed-by: Hudson CI