ctf: expose stream in CtfTmfEvent
This allows one to access the stream number. It may lead to improved UX later.
The stream ID can allow analyses to know which stream an event is coming
from. If a trace has a control channel and a data channel, this will
allow the events to be easily filtered into the appropriate views.

Change-Id: Iba9de10d9f1f633deed6df81c0d849acf5a3aceb
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/62779
Reviewed-by: Hudson CI
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>