analysis: Make SystemCallLatencyTableAnalysisTest more stable
Even if the jobs are finished, the table is not necessarily filled yet.
The update could be pending as a event in the UI thread queue, etc. We
can wait until cells are filled instead.

Also export the package so that it can be used elsewhere, (alltests
stress test for example).

Change-Id: I8177c7e6b7e487d59b248e0279c6bcd8da43b515
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/62959
Reviewed-by: Hudson CI
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>