tmf: Make font depend on item height in time graph
The font used for drawing item names and state labels is now dependent
on each item's height.

The time graph combo no longer has to adjust the time graph control's
font when zooming vertically.

A map of fonts per height is used.

The item state rectangles are adjusted to use less margin height when
the item height is decreased.

Change-Id: I1838855d864e76646bc8e387682ca0cb3c28c418
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/62980
Reviewed-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>