tmf: Fix failure in StandardImportGzipTraceTest related to test ordering
If TmfAlignTimeAxisTest ran before, and the window size was small
enough, it was possible for the test to be running in an "Align 1"
perspective with insufficient height for the events table. Because of
that, the expected content in the table could not be found. The tests in
TmfAlignTimeAxisTest switch to generated perspectives without returning
to a "normal" perspective. It makes sense that the tests "cleanup" by
returning to a more normal perspective like the Tracing perspective.

Also switch to Tracing perspective explicitly in
StandardImportGzipTraceTest which is more consistent with the other
tests and is somewhat less error prone.

Change-Id: I0d28af39266bdcdc81e765b5522c92a4a8085602
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/63081
Reviewed-by: Hudson CI
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>
(cherry picked from commit 3bd360ceec5bd7791f6476cab5f99dc665f932ae)
Reviewed-on: https://git.eclipse.org/r/63107