ss: Bug 485463: Incorrect parent sequence number in HTNode
When creating a new 'latest branch' in HistoryTree.addNewRootNode(), the
new nodes' parent sequence number is incorrectly set to their parent's
parent sequence number instead of their parent's sequence number.

The toString() implementation of HTNode used for debugging purposes is
augmented to show the sequence number of parent and children nodes. This
helps in debugging problems such as this one.

Change-Id: Ie6ec689bb28c72eda612d4279b6d27ead7ecc42c
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>
Reviewed-on: https://git.eclipse.org/r/63898
Reviewed-by: Hudson CI