ss: Bug 484776: Incorrect end time in HistoryTree
The ThreadedHistoryTreeBackend now keeps track of its own end time that
includes the end time of any interval that is in the interval queue.
This ensures that the history tree is closed with the correct end time.

The HistoryTree no longer updates its end time when initializing new
empty nodes. This ensures that the end time is never greater than the
real end time.

Change-Id: I8ff41d14f54ec1a33b8e22e81a21871f0cbf7ea0
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/64015
Reviewed-by: Hudson CI
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Tested-by: Genevieve Bastien <gbastien+lttng@versatic.net>