vm: change private method to take a List instead of ArrayList
Declarations should use Java collection interfaces such as "List"
rather than specific implementation classes such as "LinkedList"

The purpose of the Java Collections API is to provide a well defined
hierarchy of interfaces in order to hide implementation details.

Implementing classes must be used to instantiate new collections,
but the result of an instantiation should ideally be stored in a
variable whose type is a Java Collection interface.

Change-Id: I380287ab4b425167e6378f1b0c4d9aa1af4a8b02
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/64943
Reviewed-by: Hudson CI
Reviewed-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Tested-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>