linux: make KernelStateProvider handle aggregate prev_states of sched_switch
Use an if statement instead of a switch to check individual bits in the
prev_state bitmask.

fixes bug 486515

An example would be any state of EXIT_TRACE where it is two bits,
EXIT_ZOMBIE and EXIT_DEAD that are active. In the previous implementation
the state would be UNKNOWN, but now it would correctly show DEAD (null).

Change-Id: Ia9372e4d196e3a87a1e902753803a87ecafb3b33
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/65139
Reviewed-by: Hudson CI
Reviewed-by: Francis Giraldeau <francis.giraldeau@gmail.com>