analysis: add per-cpu cpu usage to analysis
Query with a set of desired cpus, or an empty set for everything.
This allows the analyses to filter on a per-cpu basis. The cpu
analysis now can be queried as follows:

This example has 4 cpus named, 0 to 3

module.getCpuUsageInRange( {}, tStart, tEnd) will get all cpus.

module.getCpuUsageInRange( {0,1,2,3}, tStart, tEnd) returns the same.

module.getCpuUsageInRange( {0}, tStart, tEnd) will return a subset
of the previous queries, affecting only CPU 0.

Note that this affects the totals too, the total is calculated on the fly.

Change-Id: Ie6a606763f4331aeef06b2916a7f98d9c73fc9d5
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/65418
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Tested-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Reviewed-by: Hudson CI