tmf: Fix TestImportExportPackageWizard on Eclipse 4.6M5
ArchiveFileExportOperation doesn't include linked resources by
default anymore, we have to call setIncludeLinkedResources(true).
Because of that, the exported package did not contain the trace
files. The method doesn't exist in Eclipse 4.5 so we have to use
reflection for the time that we want to support both versions of
Eclipse.

Change-Id: Ie7216e81fd3004e362e4200946e6ce81f81aa676
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/65503
Reviewed-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>