tmf: Keep focused item aligned on vertical zoom
When doing a vertical zoom with the keyboard, if the selected entry is
visible, make sure that it remains at the same vertical position.

When doing a vertical zoom with the mouse wheel, make sure that the
entry under the mouse cursor remains at the same vertical position.

When repeatedly doing a vertical zoom within a short amount of time, use
the initial event's entry and position for the subsequent alignments.

Fix the method getItemIndexAtY() to find items that are above the
visible bounds.

Change-Id: I803229c5cf5295b855d4615a6756751158b78346
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/65605
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>