linux: introduce execution contexts in resources view
The original implementation made sense only one single core systems. An
IRQ and softirq would block the whole system and not just a single core.

This patch adds the notion of execution contexts to the interrupts. The
interrupts are moved under the CPU to show the true execution stack.
Then the model is a much truer representation of the kernel.

In order to maintain aggregate information in the resources view (total
IRQs and such), the notion of AggregateEntries and AggregateIterators
are introduced.

This allows better understanding of the current state of the system as
the resources view now shows what the state of each processor is.

Expect a _SLIGHT_ performance degradation during the building of the
state system as an additional lookup is required for each interrupt
event.

This updates the state system model so the testvalues are updated and
the generator is updated to work again.

Starts to fix bugs: bug 381497 and bug 481855

Change-Id: I2f0f2093344f51caaf08500181cf90400a081f7c
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/65609
Reviewed-by: Hudson CI