ctf: Increase time-out of CtfTestSuiteTest
This has been timing out intermittently and I haven't been able to
reproduce it in a way that I can diagnose the problem further. A theory
is that the build machine just gets slow sometimes (it is shared between
many projects). Increasing the time-out shouldn't hurt and if it still
happens, it will be even more suspicious.

Bug: 486209
Change-Id: Ib71f8313b02ce6152bb2eb119bc9ffa11fd44d68
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/65851
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>