analysis: Bug 487388: Fix intermittent SWTBot test failure
The intermittent failure is due to the issue with the process having
the wrong parent which patch tried to fix:
https://git.eclipse.org/r/#/c/65024/

In the filter dialog the number of processes checked varies due the
issue above. So, when checking the value it might fail from time to
time. However, this is irrelevant to this test class and doesn't need
to be verified. The Filter dialog is tested elsewhere and doesn't need
to be verified in this test (see ControlFlowViewTest).

Change-Id: Ib48ff75e8ffc029f1e814e163369d64ed80916c7
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/66124
Reviewed-by: Hudson CI
Reviewed-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Tested-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>