ctf: Ignore invalid clock description elements
The clock "offset" element is supposed to be signed integer constant[1].
LTTng uses 64-bit integers for this, however there is a bug with current
versions where this can be set to values above 2^63.

Right now Trace Compass refuses to open the trace and reports the
invalid value as a CTF error. We could accomodate and still load these
traces, ignoring the invalid definitions.

[1] This is not explicitly mentionned in the spec, but it was discuted
    with the spec's authors, and it should be updated in the future.

Change-Id: I44fc4f4b90c1c9f046c03d83d63cea2fee8bb57f
Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>
Reviewed-on: https://git.eclipse.org/r/66159
Reviewed-by: Hudson CI
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>