tmf: Fix some tests not working in different time zones
Also print the timezone at the start of test to help troubleshoot
if it happens on a server.

Reviewed-on: https://git.eclipse.org/r/66579

This is a combination of 2 commits because the first one introduced
a regression and the second fixed it.

---Second commit---

tmf: Fix TestRefreshTextTrace failure because of bad time zone reset

Commit 6941941dae56 (https://git.eclipse.org/r/#/c/66579/)
changed tests to set explicitly a time zone but it doesn't
reset it properly. Because of this, TestRefreshTextTrace had
mismatching time stamps.

This can reproduced by running FilterViewerTest first then
running TestRefreshTextTrace. The easiest way is to modify
SWTBotStressTests to have this:

s.addTest(new JUnit4TestAdapter(FilterViewerTest.class));
s.addTest(new JUnit4TestAdapter(TestRefreshTextTrace.class));

To fix it, we reset the time zone to the time zone id, not the label.
The safest way to do this is to use what we decided to use in the
default preference map.

Change-Id: I35afe305ac6c81c1ccbb5c01b4031aaa2074e420
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/66723
Reviewed-on: https://git.eclipse.org/r/68881
Reviewed-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>