tmf.core: Add cpu and thread to trace context
This patch allows to see what the last selected cpu and thread were
for a given trace. The cpu and thread are stores in a LinuxTraceContext.

To do this, a default method is introduced to ITmfTrace (createTraceContext).
This method is overridden by IKernelTrace to make LinuxTraceContext instead of
the regular TmfTraceContexts. The LinuxTrace Context has the CPU and Thread ID
information. An important thing to note is that if two kernel traces have the same
hostID they both get the signal. This is by design.

Change-Id: I25f044ec05fbd69acf93009861b464506327d616
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/66667
Reviewed-by: Hudson CI
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Tested-by: Genevieve Bastien <gbastien+lttng@versatic.net>