doc: Add explicit dependency on mylyn.wikitext.core.ant
With recent versions of Mylyn Wikitext, it seems the
"org.eclipse.mylyn.wikitext.core.ant" bundle (which contains the Ant
task definitions) is not depended on by
"org.eclipse.mylyn.wikitext.core" anymore, so it does not end up
on Maven's compile classpath, and is not available for the Ant scripts.

Rectify this by adding an explicit dependency to it in the doc/ plugins.

Note that this is a compile-time dependency only, not a runtime one,
so we specify it in the build.properties file.

This should fix a current build failure with the staging target.

Change-Id: I358da1ed51754c3b355cb5e7f9a2d92e0e65f2a6
Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>
Reviewed-on: https://git.eclipse.org/r/66773
Reviewed-by: Hudson CI
Reviewed-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Tested-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>