tmf.ui: introduce Line entries into the time graph view
The timegraph view could only display "blocks" like a gantt chart, this
patch adds the ability to display multiple lines per time graph entry.
The lines are stored in a TimeGraphLineEntry. This entry has multiple
TimeLineEvents which contain multiple values and an X time.

It is absolutely necessary to have the TimeLineEvents in chronological
order or the line will go back in time.

Does not handle lines leading from/to the previous/next points.

Does not handle negative values.

Change-Id: I49e30ff14257d477a3fc27fac0a481a3a0f86e16
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/66936
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: CI Bot