lttng: Add irq_softirq_raise to lttng 2.7 layout
This patch adds the definition for soft irq raises.

When LTTng 2.7 came out, it renamed softirq raise, entry and exit
events. That patch[1] was tested using the critical path analysis which
does not look up the raise event, only entry and exit. The only view
that shows raised states is the resources view, This patch fixes the
regression on that patch. Testing is only done with an LTTng 2.0 trace
therefore it fell through the cracks on that side too.

[1] e8e9291 lttng: additions definitions for lttng 2.7 and 2.8

Change-Id: Ibd9902d57a8dfb908153df257650235f369a48e0
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/66877
Reviewed-by: Hudson CI