control: Bug 487847: Fix intermittent failure of ControlViewProfileTest
On Hudson multiple instances of this test can run at the same time.
The test reads and writes to the same directory for lttng profiles.
This is not the problem. The problem is that all instances use the same
file name which is derived from the LTTng session. To fix this, each
instance needs to generate session different session names. In the
implementation the method System.nanoTime() is used for that.

Since the test uses a LTTng Tools stub where the input and output of a
command are read from a testfile the session name must not be hardcoded
in the file. The stub was update so that it's possible to inject a
session name.

Change-Id: Idab89827e4ba5e2670d3ee947a1406a5a6dd78ed
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/66943
Reviewed-by: Hudson CI
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>