control: support creating of experiments when importing a trace session
The remote import wizard that is used to import LTTng session has been
updated to to create an experiment using all successfully imported
traces. Unrecognized traces won't be added to the experiment.
Traces that are skipped during the import (name conflict) are not
added to the experiment.

The name of the experiment is by default session root directory or
session name if the session doesn't use the default session path.
The user has the possibility to change the experiment name.

Change-Id: I6a6013511237ceb308fe25bdc42ec342a13ab72d
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/67178
Reviewed-by: Hudson CI
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>