tmf: Add an attribute pool for state system analyses
This new class allows to reuse attribute quarks in a state system. It can
typically be used by analyses who want to save data in the state system
for short intervals of time. Instead of creating a number of attributes with
lots of empty space before and after, they can reuse the same. It makes for
more compact time graph views as well.

Change-Id: Ie7f32eccb02096ef62df4c479a8e2bcfcd043af6
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/67263
Reviewed-by: Hudson CI
Reviewed-by: Alexandre Montplaisir <alexmonthy@efficios.com>