tmf: Create experiment when importing traces import trace wizard
Using the standard import wizard it's now possible to create an
experiment using all successfully imported traces. Unrecognized
traces won't be added to the experiment. Traces that are skipped
during the import (name conflict) are not added to the experiment.

The name of the experiment is by default the root folder when importing
from a directory. Or the archive name when importing from archive.
The user has the possibility to change the experiment name.

The option to create an experiment is persisted so that the checkbox
is selected/deselect when reopening the import wizard. 

Change-Id: I06b15419ebd90983865c770d7a771bc5d190cc39
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/67168
Reviewed-by: Hudson CI
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>