tmf: fix regression in trace type detection
Patch 281def4 moved the trace type detection to tmf.core. With this
change the traces that validate with confidence 0 for more than one
trace type will have a trace type set that is first in the list.
Before previous patch the trace type was not set if the confidence
was 0.

Change-Id: I2fb3dbd933ae673ce54569356f190c19a7d72d49
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/67346
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-by: Hudson CI