tmf: Move icon and label text into ITmfProjectModelElement
A series of "else if (instance of ...)" checks is usually better
served by good ol' polymorphism. Move the class-aware logic present
in TmfNavigatorLabelProvider into each class and add corresponding
methods in the interface.

This aligns with the recently-added getStyler() method, which is
defined in the element classes and simply called by the label provider.

Also move everything icon-reading-related to a new class, so that
there is no circular dependencies between TmfNavigatorLabelProvider
and the project element classes.

Change-Id: I0d78dd5342b13272a42b11e7ca0619225e1beed5
Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>
Reviewed-on: https://git.eclipse.org/r/67694
Reviewed-by: Hudson CI
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Tested-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>