tmf.core: Introduce TmfTimestamp factory methods
Often times, a TmfTimestamp will be created using a scale of -9 or 0.
The result is a less efficient timestamp that takes more memory than
the nano or simple timestamp and is slower to compare. This patch makes
timestamps created always return the more efficient type.

At the moment, the factory will either create a new TmfNanoTimestamp if the
scale is NANOSECOND_SCALE, TmfSecondTimestamp if the scale is SECOND_SCALE,
or a plain old TmfTimestamp (TmfRealTimestamp).

This patch adds the following helper methods:
* create(value, scale) - create a timestamp with an arbitrary scale and value
* fromNanos(value) - create a nanosecond based timestamp (more efficient)
* fromMicros(value) - create a microsecond based timestamp
* fromMillis(value) - create a millisecond based timestamp
* fromSeconds(value) - create a second based timestamp (more efficient)

Also, TmfTimestamp is converted to an abstract class with "getScale()" and
"getValue()" needing to be implemented. This allows a timestamp like
TmfNanoTimestamp or TmfSecondTimestamp to not store an extra integer and
save 33% of their sizes.

Change-Id: I0996c9b4f626a70436b58e82446ed69d2e50af98
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/65312
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-by: Hudson CI