tmf: Fix Select Next/Previous Marker navigation
Because the markers are sorted by time, when navigating to the next or
previous marker, the marker start time should always be visible.

The action should not rely on the 'ensureVisible' implementation of the
method setSelectionRangeNotify(), as this method will ensure that the
end time of the selection will be made visible when the start time has
not changed. This causes problems if there are two markers with the same
start time but the end time of the second is out of visible range. Since
the window range is modified, the resulting marker list may no longer
include the marker that should have been next in the navigation order.

Change-Id: I434d454d7411f8a00bf58406ad7acc3128472de7
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/67643
Reviewed-by: Hudson CI