tmf.core: add exception logging to event requests
This is the first step towards being able to properly show errors
that occur during a trace read. The TmfEventThread already logs all
Exceptions, now the request will get the said exception.

It can then be accessed by "Request#getFailCause()" and the results
can either be handled accordingly or displayed.

Change-Id: I99242dbf6eed25c4138239dd465c16fee62bb752
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/67774
Reviewed-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>