Revert "Revert "ctf: Fix slow test due to expected timeout""
This reverts commit 1ae81c23b427deeec3266e4755e9122ddcd0cf81.

This almost worked before but it needed to use withMnemonic not
withLabel!

Original message:
The test was looking for a possible checkbox (in 4.6) but when it it was
not there (4.5) it had to wait 50 secs for it to time out. Instead we
can set a very short timeout because it will instantly either be there
or not.

Change-Id: I84f6db70a323611b1adb93c6d9b55801bbeb7444
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/67782
Reviewed-by: Hudson CI
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>