tmf: make checkpoint indexer return a failed status if the request fails.
Trace compass uses the TmfIndexingJob for every trace, propagating errors
in the trace reading could be done here. Before this patch, running the
job would always return Status#OK_STATUS. Now, if there is an error that
occured (the request fails) it will propagate the error instead of
swallowing it.

Change-Id: I7247de99af4c32c6747f0100d1c07275783264fd
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/67918
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Hudson CI